pyproject.toml
[tool.poetry]
name = "auth-service"
version = "0.1.0"
description = "Scholar Spark Authentication Service"
authors = ["Pouya Ataei <pouya.ataei.7@gmail.com>"]
readme = "README.md"
packages = [{include = "app"}]  

[tool.poetry.dependencies]
python = "^3.8"
fastapi = "^0.104.0"
uvicorn = "^0.23.2"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-multipart = "^0.0.6"
psycopg2-binary = "^2.9.9"
pydantic = "^2.4.2"
pydantic-settings = "^2.0.3"
opentelemetry-api = "^1.20.0"
opentelemetry-sdk = "^1.20.0"
opentelemetry-instrumentation-fastapi = "^0.41b0"
email-validator = "^2.2.0"
scholar-spark-observability = "^0.8.0"
httpx = "^0.28.1"
redis = "^5.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"